requirements.txt
asgiref==3.3.1
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==2.2.10
django-adminlte==1.2
django-adminlte-3==0.1.6
django-adminlte-ui==1.5.1
django-adminlte3==0.1.6
django-ckeditor==6.0.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-treebeard==4.3
djangorestframework==3.12.2
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
Pillow==8.1.0
psycopg2==2.8.6
pycodestyle==2.6.0
python-dateutil==1.5
python-decouple==3.4
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
whitenoise==5.2.0
